+++ src/constants.ts
@@ -6,6 +6,7 @@ export class Constants {
     public static readonly CONFIGURATION_CHAINS = "cosmywasmy.chains";
     public static readonly CONFIGURATION_CONTRACT_SORT_ORDER = "cosmywasmy.contractSortOrder";
     public static readonly CONFIGURATION_COSMWASM_RESPONSE_VIEW = "cosmywasmy.cosmwasmResponseView";
+    public static readonly CONFIGURATION_QUERY_HISTORY = "cosmywasmy.maxQueryStored";
     
     public static readonly VIEWS_ACCOUNT = "account";
     public static readonly VIEWS_CONTRACT = "contract";